content/canvas/test/webgl/conformance/extensions/webgl-experimental-compressed-textures.html

<!--
Copyright (c) 2011 The Chromium Authors. All rights reserved.
Use of this source code is governed by a BSD-style license that can be
found in the LICENSE file.
 -->
<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>WebGL WEBGL_EXPERIMENTAL_compressed_textures Conformance Tests</title>
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<style>
img {
 border: 1px solid black;
 margin-right: 1em;
}
.testimages {
}

.testimages br {
  clear: both;
}

.testimages > div {
  float: left;
  margin: 1em;
}
</style>
<script src="../../resources/desktop-gl-constants.js" type="text/javascript"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/webgl-test.js"></script>
<script src="../resources/webgl-test-utils.js"></script>
<script src="compressed-textures/4x4.rgb.dxt1.js"></script>
<script src="compressed-textures/4x4.rgba.dxt1.js"></script>
<script src="compressed-textures/4x4.rgba.dxt5.js"></script>
<script src="compressed-textures/8x8.rgb.dxt1.js"></script>
<script src="compressed-textures/8x8.rgba.dxt1.js"></script>
<script src="compressed-textures/8x8.rgba.dxt5.js"></script>
<script src="compressed-textures/8x8.rgba.pvrtc4bpp.js"></script>
</head>
<body>
<div id="description"></div>
<canvas id="canvas" width="8" height="8" style="width: 8px; height: 8px;"></canvas>
<div id="console"></div>
<!-- Shaders for testing standard derivatives -->
<script>
description("This test verifies the functionality of the OES_vertex_array_object extension, if it is available.");

debug("");

var wtu = WebGLTestUtils;
var canvas = document.getElementById("canvas");
var gl = wtu.create3DContext(canvas, {antialias: false});
var program = wtu.setupTexturedQuad(gl);
var ext = null;
var vao = null;
var validFormats = {
  COMPRESSED_RGB_S3TC_DXT1_EXT        : 0x83F0,
  COMPRESSED_RGBA_S3TC_DXT1_EXT       : 0x83F1,
  COMPRESSED_RGBA_S3TC_DXT5_EXT       : 0x83F3,
  ETC1_RGB8_OES                       : 0x8D64,
  COMPRESSED_RGB_PVRTC_4BPPV1_IMG     : 0x8C00,
  COMPRESSED_RGBA_PVRTC_4BPPV1_IMG    : 0x8C02
};
var name;
var supportedFormats;

if (!gl) {
  testFailed("WebGL context does not exist");
} else {
  testPassed("WebGL context exists");

  // Run tests with extension disabled
  runTestDisabled();

  // Query the extension and store globally so shouldBe can access it
  ext = gl.getExtension("WEBGL_EXPERIMENTAL_compressed_textures");
  if (!ext) {
    testPassed("No WEBGL_EXPERIMENTAL_compressed_textures support -- this is legal");
    runSupportedTest(false);
  } else {
    testPassed("Successfully enabled WEBGL_EXPERIMENTAL_compressed_textures extension");

    runSupportedTest(true);
    runTestExtension();
  }
}

function runSupportedTest(extensionEnabled) {
  var supported = gl.getSupportedExtensions();
  if (supported.indexOf("WEBGL_EXPERIMENTAL_compressed_textures") >= 0) {
    if (extensionEnabled) {
      testPassed("WEBGL_EXPERIMENTAL_compressed_textures listed as supported and getExtension succeeded");
    } else {
      testFailed("WEBGL_EXPERIMENTAL_compressed_textures listed as supported but getExtension failed");
    }
  } else {
    if (extensionEnabled) {
      testFailed("WEBGL_EXPERIMENTAL_compressed_textures not listed as supported but getExtension succeeded");
    } else {
      testPassed("WEBGL_EXPERIMENTAL_compressed_textures not listed as supported and getExtension failed -- this is legal");
    }
  }
}


function runTestDisabled() {
  debug("Testing binding enum with extension disabled");

  // Default value is null
  if (gl.getParameter(gl.COMPRESSED_TEXTURE_FORMATS) === null) {
    testPassed("Default value of COMPRESSED_TEXTURE_FORMATS is null");
  } else {
    testFailed("Default value of COMPRESSED_TEXTURE_FORMATS is not null");
  }
}

function checkIsValidFormat(format) {
  for (var name in validFormats) {
    if (format == validFormats[name]) {
      testPassed("supported format " + formatToString(format) + " is valid");
      return;
    }
  }
  testFailed("supported format " + formatToString(format) + " is not valid");
}

function formatToString(format) {
  for (var p in ext) {
    if (ext[p] == format) {
      return p;
    }
  }
  return "0x" + format.toString(16);
}

function runTestExtension() {
  debug("Testing WEBGL_EXPERIMENTAL_compressed_textures");

  // check that all format enums exist.
  for (name in validFormats) {
    var expected = "0x" + validFormats[name].toString(16);
    var actual = "ext['" + name + "']";
    shouldBe(actual, expected);
  }

  supportedFormats = gl.getParameter(gl.COMPRESSED_TEXTURE_FORMATS);
  // Even if no formats are supported this should return an array.
  shouldBeTrue("supportedFormats.length !== undefined");

  // check that each format is an allowed format
  for (var ii = 0; ii < supportedFormats.length; ++ii) {
    checkIsValidFormat(supportedFormats[ii]);
  }

  // Test each format
  for (var ii = 0; ii < supportedFormats.length; ++ii) {
    switch (supportedFormats[ii]) {
    case ext.COMPRESSED_RGB_S3TC_DXT1_EXT:
      testDXT1_RGB();
      break;
    case ext.COMPRESSED_RGBA_S3TC_DXT1_EXT:
      testDXT1_RGBA();
      break;
    case ext.COMPRESSED_RGBA_S3TC_DXT5_EXT:
      testDXT5_RGBA();
      break;
    case ext.ETC1_RGB8_OES:
      testETC1_RGB8();
      break;
    case ext.COMPRESSED_RGB_PVRTC_4BPPV1_IMG:
      testPVRTC_RGB_4BPPV1();
      break;
    case ext.COMPRESSED_RGBA_PVRTC_4BPPV1_IMG:
      testPVRTC_RGBA_4BPPV1();
      break;
    }
  }
}

function testDXT1_RGB() {
  var tests = [
    { width: 4,
      height: 4,
      channels: 3,
      data: img_4x4_rgb_dxt1,
      format: ext.COMPRESSED_RGB_S3TC_DXT1_EXT
    },
    { width: 8,
      height: 8,
      channels: 3,
      data: img_8x8_rgb_dxt1,
      format: ext.COMPRESSED_RGB_S3TC_DXT1_EXT
    }
  ];
  testDXTTextures(tests);
}

function testDXT1_RGBA() {
  var tests = [
    { width: 4,
      height: 4,
      channels: 4,
      data: img_4x4_rgba_dxt1,
      format: ext.COMPRESSED_RGBA_S3TC_DXT1_EXT
    },
    { width: 8,
      height: 8,
      channels: 4,
      data: img_8x8_rgba_dxt1,
      format: ext.COMPRESSED_RGBA_S3TC_DXT1_EXT
    }
  ];
  testDXTTextures(tests);
}

function testDXT5_RGBA() {
  var tests = [
    { width: 4,
      height: 4,
      channels: 4,
      data: img_4x4_rgba_dxt5,
      format: ext.COMPRESSED_RGBA_S3TC_DXT5_EXT
    },
    { width: 8,
      height: 8,
      channels: 4,
      data: img_8x8_rgba_dxt5,
      format: ext.COMPRESSED_RGBA_S3TC_DXT5_EXT
    }
  ];
  testDXTTextures(tests);
}

function testETC1_RGB8() {
  testFailed("ETC1 test not yet implemented");
}

function testPVRTC_RGB_4BPPV1() {
  var tests = [
    { width: 8,
      height: 8,
      channels: 4,
      data: img_8x8_rgba_pvrtc4bpp,
      format: ext.COMPRESSED_RGB_PVRTC_4BPPV1_IMG
    }
  ];
  testPVRTCTextures(tests);
}

function testPVRTC_RGB_4BPPV1() {
  var tests = [
    { width: 8,
      height: 8,
      channels: 4,
      data: img_8x8_rgba_pvrtc4bpp,
      format: ext.COMPRESSED_RGBA_PVRTC_4BPPV1_IMG
    }
  ];
  testPVRTCTextures(tests);
}

function testDXTTextures(tests) {
  debug("<hr/>");
  for (var ii = 0; ii < tests.length; ++ii) {
    testDXTTexture(tests[ii]);
  }
}

function uncompressDXTBlock(
    destBuffer, destX, destY, destWidth, src, srcOffset, format) {
  function make565(src, offset) {
    return src[offset + 0] + src[offset + 1] * 256;
  }
  function make8888From565(c) {
    return [
        Math.floor(((c >> 11) & 0x1F) * 255 / 31),
        Math.floor(((c >>  5) & 0x3F) * 255 / 63),
        Math.floor(((c >>  0) & 0x1F) * 255 / 31),
        255
      ];
  }
  function mix(mult, c0, c1, div) {
    var r = [];
    for (var ii = 0; ii < c0.length; ++ii) {
      r[ii] = Math.floor((c0[ii] * mult + c1[ii]) / div);
    }
    return r;
  }
  var colorOffset =
      srcOffset + ((format == ext.COMPRESSED_RGBA_S3TC_DXT5_EXT) ? 8 : 0);
  var color0 = make565(src, colorOffset + 0);
  var color1 = make565(src, colorOffset + 2);
  var c0gtc1 = color0 > color1 || format == ext.COMPRESSED_RGBA_S3TC_DXT5_EXT;
  var rgba0 = make8888From565(color0);
  var rgba1 = make8888From565(color1);
  var colors = [
      rgba0,
      rgba1,
      c0gtc1 ? mix(2, rgba0, rgba1, 3) : mix(1, rgba0, rgba1, 2),
      c0gtc1 ? mix(2, rgba1, rgba0, 3) : [0, 0, 0, 255]
    ];

  // yea I know there is a lot of math in this inner loop.
  // so sue me.
  for (var yy = 0; yy < 4; ++yy) {
    var pixels = src[colorOffset + 4 + yy];
    for (var xx = 0; xx < 4; ++xx) {
      var dstOff = ((destY + yy) * destWidth + destX + xx) * 4;
      var code = (pixels >> (xx * 2)) & 0x3;
      var srcColor = colors[code];
      var alpha;
      switch (format) {
      case ext.COMPRESSED_RGB_S3TC_DXT1_EXT:
        alpha = 255;
        break;
      case ext.COMPRESSED_RGBA_S3TC_DXT1_EXT:
        alpha = (code == 3 && !c0gtc1) ? 0 : 255;
        break;
      case ext.COMPRESSED_RGBA_S3TC_DXT5_EXT:
        {
          var alpha0 = src[srcOffset + 0];
          var alpha1 = src[srcOffset + 1];
          var alphaOff = Math.floor(yy / 2) * 3 + 2;
          var alphaBits =
             src[srcOffset + alphaOff + 0] +
             src[srcOffset + alphaOff + 1] * 256 +
             src[srcOffset + alphaOff + 2] * 65536;
          var alphaShift = (yy % 2) * 12 + xx * 3;
          var alphaCode = (alphaBits >> alphaShift) & 0x7;
          if (alpha0 > alpha1) {
            switch (alphaCode) {
            case 0:
              alpha = alpha0;
              break;
            case 1:
              alpha = alpha1;
              break;
            default:
              alpha = ((8 - alphaCode) * alpha0 + (alphaCode - 1) * alpha1) / 7;
              break;
            }
          } else {
            switch (alphaCode) {
            case 0:
              alpha = alpha0;
              break;
            case 1:
              alpha = alpha1;
              break;
            case 6:
              alpha = 0;
              break;
            case 7:
              alpha = 255;
              break;
            default:
              alpha = ((6 - alphaCode) * alpha0 + (alphaCode - 1) * alpha1) / 5;
              break;
            }
          }
        }
        break;
      default:
        throw "bad format";
      }
      destBuffer[dstOff + 0] = srcColor[0];
      destBuffer[dstOff + 1] = srcColor[1];
      destBuffer[dstOff + 2] = srcColor[2];
      destBuffer[dstOff + 3] = alpha;
    }
  }
}

function uncompressDXT(width, height, data, format) {
  if (width % 4 || height % 4) throw "bad width or height";

  var dest = new Uint8Array(width * height * 4);
  var blocksAcross = width / 4;
  var blocksDown = height / 4;
  var blockSize = (format == ext.COMPRESSED_RGBA_S3TC_DXT5_EXT ? 16 : 8);
  for (var yy = 0; yy < blocksDown; ++yy) {
    for (var xx = 0; xx < blocksAcross; ++xx) {
      uncompressDXTBlock(
        dest, xx * 4, yy * 4, width, data,
        (yy * blocksAcross + xx) * blockSize, format);
    }
  }
  return dest;
}

function testDXTTexture(test) {
  var uncompressedData = uncompressDXT(
      test.width, test.height, test.data, test.format);

  var inner = 4; //test.width == 4 ? 4 : 1
  for (var ii = 0; ii < inner; ++ii) {
    var width = test.width - ii;
    var height = test.height - ii;
    canvas.width = width;
    canvas.height = height;
    gl.viewport(0, 0, width, height);
    debug("testing " + formatToString(test.format) + " " +
          width + "x" + height);
    var tex = gl.createTexture();
    gl.bindTexture(gl.TEXTURE_2D, tex);
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
    ext.compressedTexImage2D(
        gl.TEXTURE_2D, 0, test.format, width, height, 0,
        new Uint8Array(test.data));
    glErrorShouldBe(gl, gl.NO_ERROR, "uploading compressed texture");
    wtu.drawQuad(gl);
    compareRect(
        width, height, test.channels, test.width, test.height, uncompressedData,
        test.data, test.format);
  }

  gl.compressedTexImage2D(
      gl.TEXTURE_2D, 0, test.format, width + 1, height, 0,
      new Uint8Array(test.data));
  glErrorShouldBe(
      gl, gl.INVALID_OPERATION, "data size does not match dimensions");


  // TODO: test compressedTexSubImage2D
  // TODO: test invalid width, height, xoffset, yoffset
}

function insertImg(element, caption, img) {
  var div = document.createElement("div");
  div.appendChild(img);
  var label = document.createElement("div");
  label.appendChild(document.createTextNode(caption));
  div.appendChild(label);
  element.appendChild(div);
}

function makeImage(imageWidth, imageHeight, dataWidth, data, alpha) {
  var scale = 8;
  var c = document.createElement("canvas");
  c.width = imageWidth * scale;
  c.height = imageHeight * scale;
  var ctx = c.getContext("2d");
  for (var yy = 0; yy < imageWidth; ++yy) {
    for (var xx = 0; xx < imageHeight; ++xx) {
      var offset = (yy * dataWidth + xx) * 4;
      ctx.fillStyle = "rgba(" +
          data[offset + 0] + "," +
          data[offset + 1] + "," +
          data[offset + 2] + "," +
          (alpha ? data[offset + 3] / 255 : 1) + ")";
      ctx.fillRect(xx * scale, yy * scale, scale, scale);
    }
  }
  var img = document.createElement("img");
  img.src = c.toDataURL();
  return img;
}
function compareRect(
    actualWidth, actualHeight, actualChannels,
    dataWidth, dataHeight, expectedData,
    testData, testFormat) {
  var actual = new Uint8Array(actualWidth * actualHeight * 4);
  gl.readPixels(
      0, 0, actualWidth, actualHeight, gl.RGBA, gl.UNSIGNED_BYTE, actual);

  var div = document.createElement("div");
  div.className = "testimages";
  insertImg(div, "expected", makeImage(
      actualWidth, actualHeight, dataWidth, expectedData,
      actualChannels == 4));
  insertImg(div, "actual", makeImage(
      actualWidth, actualHeight, actualWidth, actual,
      actualChannels == 4));
  div.appendChild(document.createElement('br'));
  document.getElementById("console").appendChild(div);

  var failed = false;
  for (var yy = 0; yy < actualHeight; ++yy) {
    for (var xx = 0; xx < actualWidth; ++xx) {
      var actualOffset = (yy * actualWidth + xx) * 4;
      var expectedOffset = (yy * dataWidth + xx) * 4;
      var expected = [
          expectedData[expectedOffset + 0],
          expectedData[expectedOffset + 1],
          expectedData[expectedOffset + 2],
          (actualChannels == 3 ? 255 : expectedData[expectedOffset + 3])
      ];
      for (var jj = 0; jj < 4; ++jj) {
        if (actual[actualOffset + jj] != expected[jj]) {
          failed = true;
          var was = actual[actualOffset + 0].toString();
          for (j = 1; j < 4; ++j) {
            was += "," + actual[actualOffset + j];
          }
          testFailed('at (' + xx + ', ' + yy +
                     ') expected: ' + expected + ' was ' + was);
        }
      }
    }
  }
  if (!failed) {
    testPassed("texture rendered correctly");
  }
}

function testPVRTCTextures() {
  testFailed("PVRTC test not yet implemented");
}

debug("");
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>

</body>
</html>
